경고 제거

+++ Project/Client/AnimationTool.cpp
@@ -202,8 +202,8 @@ void AnimationTool::render_update()
     ImVec4 tint_col = use_text_color_for_tint ? ImGui::GetStyleColorVec4(ImGuiCol_Text) : ImVec4(1.0f, 1.0f, 1.0f, 1.0f); // No tint
     ImVec4 border_col = ImGui::GetStyleColorVec4(ImGuiCol_Border);
     int width = 800;
-    int height = (float)m_Atlas.Get()->GetHeight() / m_Atlas.Get()->GetWidth() * width;
-    m_RealImgSize = ImVec2(m_Atlas.Get()->GetWidth(), m_Atlas.Get()->GetHeight());
+    int height = int((float)m_Atlas.Get()->GetHeight() / m_Atlas.Get()->GetWidth() * width);
+    m_RealImgSize = ImVec2((float)m_Atlas.Get()->GetWidth(), (float)m_Atlas.Get()->GetHeight());
     m_ImgSize = { (float)width, (float)height };
 
 
@@ -291,7 +291,7 @@ void AnimationTool::render_update()
         ButtonTitle(str.c_str());
         str = "##" + str;
 
-        int lt[2] = { (float)m_vecLeftTops[i].x, (float)m_vecLeftTops[i].y };
+        int lt[2] = { (int)m_vecLeftTops[i].x, (int)m_vecLeftTops[i].y };
         ImGui::InputInt2(str.c_str(), lt);
 
         ImGui::SameLine();
