+++ Project/Client/Inspector_init.cpp
@@ -106,7 +106,7 @@ void Inspector::CreateAssetUI()
 
 void Inspector::ResizeScriptUI(UINT _Size)
 {
-	int AddSize = _Size - m_vecScriptUI.size();
+	int AddSize = _Size - (int)m_vecScriptUI.size();
 
 	for (int i = 0; i < AddSize; ++i)
 	{
